REFACTOR: Extract topic level assign menu code into separate file
